Make includable through jitpack
Requires hardcoding the name of the repo in ExternalVersionExtension, because
jitpack overrides groupId:

https://github.com/jitpack/jitpack.io/issues/1029

Furthermore, for multi-module projects there can only be one groupId:

https://github.com/jitpack/jitpack.io/issues/2872

So changed everything to be in *.version (i.e. not in *.plugin). Maven
complains that maven-*-plugin should be in groupId org.apache.maven.plugins,
but works anyway.